#914: WIP Adds test for indexing articles
Adds test for indexing article using localNodes from elastic4s-embedded.
Also makes use of joda-time for searchablearticles.
