+++ src/main/scala/no/ndla/searchapi/model/domain/article/Content.scala
@@ -7,8 +7,7 @@
 
 package no.ndla.searchapi.model.domain.article
 
-import java.util.Date
-
+import org.joda.time.DateTime
 import org.json4s.FieldSerializer
 
 sealed trait Content {
@@ -26,8 +25,8 @@ case class Article(id: Option[Long],
                    introduction: Seq[ArticleIntroduction],
                    metaDescription: Seq[ArticleMetaDescription],
                    metaImageId: Option[String],
-                   created: Date,
-                   updated: Date,
+                   created: DateTime,
+                   updated: DateTime,
                    updatedBy: String,
                    articleType: String) extends Content
 
@@ -54,8 +53,8 @@ case class Concept(id: Option[Long],
                    title: Seq[ConceptTitle],
                    content: Seq[ConceptContent],
                    copyright: Option[Copyright],
-                   created: Date,
-                   updated: Date) extends Content
+                   created: DateTime,
+                   updated: DateTime) extends Content
 
 object Concept {
   implicit val formats = org.json4s.DefaultFormats
