--- src/test/java/no/ndla/tag/IntegrationTest.java
@@ -1,14 +1,0 @@
-package no.ndla.tag;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@org.scalatest.TagAnnotation
-@Retention(RUNTIME)
-@Target({METHOD, TYPE})
-public @interface IntegrationTest {
-}
