+++ src/test/scala/no/ndla/searchapi/service/search/ArticleSearchServiceTest.scala
@@ -8,14 +8,15 @@
 
 package no.ndla.searchapi.service.search
 
+import com.sksamuel.elastic4s.embedded.LocalNode
 import no.ndla.searchapi.integration.Elastic4sClientFactory
 import no.ndla.searchapi.model.domain.{Language, Sort}
 import no.ndla.searchapi.model.domain.article._
 import no.ndla.searchapi.{SearchApiProperties, TestData, TestEnvironment, UnitSuite}
+import no.ndla.searchapi.TestData._
 import no.ndla.searchapi.SearchApiProperties.DefaultPageSize
 import com.sksamuel.elastic4s.http.ElasticDsl._
 import no.ndla.searchapi.model.taxonomy._
-import no.ndla.tag.IntegrationTest
 import org.joda.time.DateTime
 import org.mockito
 import org.mockito.Matchers._
@@ -23,132 +24,19 @@ import org.mockito.Mockito._
 
 import scala.util.Success
 
-@IntegrationTest
 class ArticleSearchServiceTest extends UnitSuite with TestEnvironment {
+  val esPort = 9202
+  val localNodeSettings: Map[String, String] = LocalNode.requiredSettings(this.getClass.getName, s"/tmp/${this.getClass.getName}") + ("http.port" -> s"$esPort")
+  val localNode = LocalNode(localNodeSettings)
 
-  val esPort = 9200
-
-  override val e4sClient = Elastic4sClientFactory.getClient(searchServer = s"http://localhost:$esPort")
+  override val e4sClient = Elastic4sClientFactory.getClient(searchServer = s"elasticsearch://${localNode.ipAndPort}")
 
   override val articleSearchService = new ArticleSearchService
   override val articleIndexService = new ArticleIndexService
   override val converterService = new ConverterService
   override val searchConverterService = new SearchConverterService
 
-  val byNcSa = Copyright("by-nc-sa", "Gotham City", List(Author("Forfatter", "DC Comics")), List(), List(), None, None, None)
-  val publicDomain = Copyright("publicdomain", "Metropolis", List(Author("Forfatter", "Bruce Wayne")), List(), List(), None, None, None)
-  val copyrighted = Copyright("copyrighted", "New York", List(Author("Forfatter", "Clark Kent")), List(), List(), None, None, None)
-
-  val today = DateTime.now()
-
-  val article1 = TestData.sampleArticleWithByNcSa.copy(
-    id = Option(1),
-    title = List(ArticleTitle("Batmen er på vift med en bil", "nb")),
-    introduction = List(ArticleIntroduction("Batmen", "nb")),
-    content = List(ArticleContent("Bilde av en <strong>bil</strong> flaggermusmann som vifter med vingene <em>bil</em>.", "nb")),
-    tags = List(ArticleTag(List("fugl"), "nb")),
-    created = today.minusDays(4).toDate,
-    updated = today.minusDays(3).toDate)
-  val article2 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(2),
-    title = List(ArticleTitle("Pingvinen er ute og går", "nb")),
-    introduction = List(ArticleIntroduction("Pingvinen", "nb")),
-    content = List(ArticleContent("<p>Bilde av en</p><p> en <em>pingvin</em> som vagger borover en gate</p>", "nb")),
-    tags = List(ArticleTag(List("fugl"), "nb")),
-    created = today.minusDays(4).toDate,
-    updated = today.minusDays(2).toDate)
-  val article3 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(3),
-    title = List(ArticleTitle("Donald Duck kjører bil", "nb")),
-    introduction = List(ArticleIntroduction("Donald Duck", "nb")),
-    content = List(ArticleContent("<p>Bilde av en en and</p><p> som <strong>kjører</strong> en rød bil.</p>", "nb")),
-    tags = List(ArticleTag(List("and"), "nb")),
-    created = today.minusDays(4).toDate,
-    updated = today.minusDays(1).toDate
-  )
-  val article4 = TestData.sampleArticleWithCopyrighted.copy(
-    id = Option(4),
-    title = List(ArticleTitle("Superman er ute og flyr", "nb")),
-    introduction = List(ArticleIntroduction("Superman", "nb")),
-    content = List(ArticleContent("<p>Bilde av en flygende mann</p><p> som <strong>har</strong> superkrefter.</p>", "nb")),
-    tags = List(ArticleTag(List("supermann"), "nb")),
-    created = today.minusDays(4).toDate,
-    updated = today.toDate
-  )
-  val article5 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(5),
-    title = List(ArticleTitle("Hulken løfter biler", "nb")),
-    introduction = List(ArticleIntroduction("Hulken", "nb")),
-    content = List(ArticleContent("<p>Bilde av hulk</p><p> som <strong>løfter</strong> en rød bil.</p>", "nb")),
-    tags = List(ArticleTag(List("hulk"), "nb")),
-    created = today.minusDays(40).toDate,
-    updated = today.minusDays(35).toDate
-  )
-  val article6 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(6),
-    title = List(ArticleTitle("Loke og Tor prøver å fange midgaardsormen", "nb")),
-    introduction = List(ArticleIntroduction("Loke og Tor", "nb")),
-    content = List(ArticleContent("<p>Bilde av <em>Loke</em> og <em>Tor</em></p><p> som <strong>fisker</strong> fra Naglfar.</p>", "nb")),
-    tags = List(ArticleTag(List("Loke", "Tor", "Naglfar"), "nb")),
-    created = today.minusDays(30).toDate,
-    updated = today.minusDays(25).toDate
-  )
-  val article7 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(7),
-    title = List(ArticleTitle("Yggdrasil livets tre", "nb")),
-    introduction = List(ArticleIntroduction("Yggdrasil", "nb")),
-    content = List(ArticleContent("<p>Bilde av <em>Yggdrasil</em> livets tre med alle dyrene som bor i det.", "nb")),
-    tags = List(ArticleTag(List("yggdrasil"), "nb")),
-    created = today.minusDays(20).toDate,
-    updated = today.minusDays(15).toDate
-  )
-  val article8 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(8),
-    title = List(ArticleTitle("Baldur har mareritt", "nb")),
-    introduction = List(ArticleIntroduction("Baldur", "nb")),
-    content = List(ArticleContent("<p>Bilde av <em>Baldurs</em> mareritt om Ragnarok.", "nb")),
-    tags = List(ArticleTag(List("baldur"), "nb")),
-    created = today.minusDays(10).toDate,
-    updated = today.minusDays(5).toDate,
-    articleType = LearningResourceType.TopicArticle.toString
-  )
-  val article9 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(9),
-    title = List(ArticleTitle("En Baldur har mareritt om Ragnarok", "nb")),
-    introduction = List(ArticleIntroduction("Baldur", "nb")),
-    content = List(ArticleContent("<p>Bilde av <em>Baldurs</em> som har  mareritt.", "nb")),
-    tags = List(ArticleTag(List("baldur"), "nb")),
-    created = today.minusDays(10).toDate,
-    updated = today.minusDays(5).toDate,
-    articleType = LearningResourceType.TopicArticle.toString
-  )
-  val article10 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(10),
-    title = List(ArticleTitle("This article is in english", "en")),
-    introduction = List(ArticleIntroduction("Engulsk", "en")),
-    content = List(ArticleContent("<p>Something something <em>english</em> What", "en")),
-    tags = List(ArticleTag(List("englando"), "en")),
-    created = today.minusDays(10).toDate,
-    updated = today.minusDays(5).toDate,
-    articleType = LearningResourceType.TopicArticle.toString
-  )
-  val article11 = TestData.sampleArticleWithPublicDomain.copy(
-    id = Option(11),
-    title = List(ArticleTitle("Katter", "nb"), ArticleTitle("Cats", "en")),
-    introduction = List(ArticleIntroduction("Katter er store", "nb"), ArticleIntroduction("Cats are big", "en")),
-    metaDescription = List(ArticleMetaDescription("hurr durr ima sheep", "en")),
-    content = List(ArticleContent("<p>Noe om en katt</p>", "nb"), ArticleContent("<p>Something about a cat</p>", "en")),
-    tags = List(ArticleTag(List("ikkehund"), "nb"), ArticleTag(List("notdog"), "en")),
-    created = today.minusDays(10).toDate,
-    updated = today.minusDays(5).toDate,
-    articleType = LearningResourceType.TopicArticle.toString
-  )
-
-  val articlesToIndex = List(
-    article1, article2, article3, article4, article5, article6, article7, article8, article9, article10, article11
-  )
-
-  override def beforeAll = {
+  override def beforeAll: Unit = {
     articleIndexService.createIndexWithName(SearchApiProperties.SearchIndexes("articles"))
 
     articlesToIndex.zipWithIndex.map{case (article: Article, index: Int) =>
@@ -196,20 +84,11 @@ class ArticleSearchServiceTest extends UnitSuite with TestEnvironment {
 
     }
 
-    blockUntil(() => articleSearchService.countDocuments == 11)
+    blockUntil(() => articleIndexService.countDocuments == 11)
   }
 
-  private def deleteIndexesThatStartWith(startsWith: String): Unit = {
-    val Success(result) = e4sClient.execute(getAliases())
-    val toDelete = result.result.mappings.filter(_._1.name.startsWith(startsWith)).map(_._1.name)
-
-    if(toDelete.nonEmpty) {
-      e4sClient.execute(deleteIndex(toDelete))
-    }
-  }
-
-  override def afterAll() = {
-    deleteIndexesThatStartWith(SearchApiProperties.SearchIndexes("articles"))
+  override def afterAll(): Unit = {
+    localNode.stop(true)
   }
 
   test("That getStartAtAndNumResults returns SEARCH_MAX_PAGE_SIZE for value greater than SEARCH_MAX_PAGE_SIZE") {
