+++ src/test/scala/no/ndla/searchapi/service/search/SearchConverterServiceTest.scala
@@ -43,7 +43,7 @@ class SearchConverterServiceTest extends UnitSuite with TestEnvironment {
     ArticleTag(Seq("the", "words"), "unknown")
   )
 
-  override def beforeAll() = {
+  override def beforeAll(): Unit = {
     when(converterService.withAgreementCopyright(any[Article])).thenAnswer((invocation: InvocationOnMock) => invocation.getArgumentAt(0, sampleArticle.getClass()))
 
     val resources = List(Resource("urn:resource:1", "Resource1", Some("urn:article:1"), "/subject:1/topic:10/resource:1"))
@@ -51,7 +51,6 @@ class SearchConverterServiceTest extends UnitSuite with TestEnvironment {
     val topicResourceConnections = List(TopicResourceConnection("urn:topic:10", "urn:resource:1", "urn:topic-resource:abc123", true, 1))
     val subjects = List(Resource("urn:subject:1", "Subject1", None, "/subject:1"))
     val subjectTopicConnections = List(SubjectTopicConnection("urn:subject:1", "urn:topic:10", "urn:subject-topic:8180abc", true, 1))
-
 
     val emptyBundle = Bundle(
       filters = List.empty,
@@ -71,13 +70,10 @@ class SearchConverterServiceTest extends UnitSuite with TestEnvironment {
   }
 
   test("That asSearchableArticle converts titles with correct language") {
-
-
     val article = TestData.sampleArticleWithByNcSa.copy(title=titles)
     val Success(searchableArticle) = searchConverterService.asSearchableArticle(article, None)
     verifyTitles(searchableArticle)
   }
-
 
   test("That asSearchable converts articles with correct language") {
     val article = TestData.sampleArticleWithByNcSa.copy(content=articles)
@@ -85,13 +81,11 @@ class SearchConverterServiceTest extends UnitSuite with TestEnvironment {
     verifyArticles(searchableArticle)
   }
 
-
   test("That asSearchable converts tags with correct language") {
     val article = TestData.sampleArticleWithByNcSa.copy(tags=articleTags)
     val Success(searchableArticle) = searchConverterService.asSearchableArticle(article, None)
     verifyTags(searchableArticle)
   }
-
 
   test("That asSearchable converts all fields with correct language") {
     val article = TestData.sampleArticleWithByNcSa.copy(title=titles, content=articles, tags=articleTags)
